tests/wpt/tests/css/css-text/parsing/tab-size-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing tab-size with invalid values</title>
<link rel="help" href="https://www.w3.org/TR/css-text-3/#propdef-tab-size">
<meta name="assert" content="tab-size supports only the grammar '<number> | <length>'.">
<meta name="assert" content="Negative tab-size values are not allowed.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("tab-size", "-10px");
test_invalid_value("tab-size", "-20");
test_invalid_value("tab-size", "30%");
test_invalid_value("tab-size", "calc(40% + 50px)");
</script>
</body>
</html>
